Extract page initialisers to dedicated file
