+++ src/hooks/usePageReducer.js
@@ -10,6 +10,7 @@ import {
   getPageActions,
   getColumns,
   getPageInfoColumns,
+  getPageInitialiser,
 } from '../pages/dataTableUtilities';
 
 import { debounce } from '../utilities/index';
@@ -37,14 +38,13 @@ import { debounce } from '../utilities/index';
  *
  * @param {String} page                   routeName for the current page.
  * @param {Object} initialState           Initial state of the reducer
- * @param {Func}   initializer            Function to generate the initial state (optional)
+ * @param {Func}   initialiser            Function to generate the initial state (optional)
  * @param {Object} pageObject             base PageObject for the page i.e. a Requisition.
  * @param {Number} debounceTimeout        Timeout period for a regular debounce
  * @param {Number} instantDebounceTimeout Timeout period for an instant debounce
  */
 export const usePageReducer = (
   initialState,
-  initializer,
   debounceTimeout = 250,
   instantDebounceTimeout = 250
 ) => {
@@ -52,9 +52,10 @@ export const usePageReducer = (
   const columns = useMemo(() => getColumns(page), []);
   const pageInfoColumns = useMemo(() => getPageInfoColumns(page), []);
   const PageActions = useMemo(() => getPageActions(page), []);
+  const initialiser = useMemo(() => getPageInitialiser(page), []);
 
   const [pageState, setPageState] = useState({
-    ...(initializer ? initializer(pageObject) : initialState),
+    ...(initialiser ? initialiser(pageObject) : initialState),
     columns,
     getPageInfoColumns: pageInfoColumns,
     PageActions,
