+++ src/pages/CustomerRequisitionsPage.js
@@ -12,28 +12,11 @@ import { View } from 'react-native';
 import { SearchBar, DataTablePageView } from '../widgets';
 import { DataTable, DataTableHeaderRow, DataTableRow } from '../widgets/DataTable';
 
-import { UIDatabase } from '../database';
-
-import { newSortDataBy } from '../utilities';
 import { usePageReducer, useNavigationFocus, useSyncListener } from '../hooks';
 import { gotoCustomerRequisition } from '../navigation/actions';
 import { getItemLayout, recordKeyExtractor } from './dataTableUtilities';
 
 import { newPageStyles } from '../globalStyles';
-
-const stateInitialiser = () => {
-  const backingData = UIDatabase.objects('ResponseRequisition');
-  const data = newSortDataBy(backingData.slice(), 'serialNumber', false);
-  return {
-    backingData,
-    data,
-    keyExtractor: recordKeyExtractor,
-    searchTerm: '',
-    filterDataKeys: ['serialNumber', 'otherStoreName.name'],
-    sortBy: 'serialNumber',
-    isAscending: false,
-  };
-};
 
 /**
  * Renders a mSupply mobile page with a list of Customer requisitions.
@@ -55,7 +38,7 @@ const stateInitialiser = () => {
  */
 export const CustomerRequisitionsPage = ({ routeName, dispatch: reduxDispatch, navigation }) => {
   const initialState = { page: routeName };
-  const [state, dispatch, debouncedDispatch] = usePageReducer(initialState, stateInitialiser);
+  const [state, dispatch, debouncedDispatch] = usePageReducer(initialState);
 
   const { data, sortBy, isAscending, searchTerm, PageActions, columns, keyExtractor } = state;
